automation: switch git hash detection to use git cherry-pick
switch git hash detection to use git cherry-pick. Using the last
non-merge commit doesn't take into consideration that the current
PR's last hash isn't necessarily the last one in the timeline --
things get merged out of order. git cherry-pick gets the commits
in the current PR/branch that aren't in master, so we use the last
hash in that list for the rpm filename.

Change-Id: Ia9c443362b62ed22b4aecbce1bda0c5e9803a7ea
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>